requirements.txt
# Direct dependencies — both verified 0 CVEs (OSV.dev, 2026-02-27)
requests>=2.32.0          # CVE-2024-35195 fixed in 2.32.0; 2.32.5 is latest
python-dotenv>=1.0.0      # no known CVEs